requirements.txt
arrow==1.3.0
asgiref==3.8.1
black==24.8.0
build==1.2.2
cassiopeia==5.1.3
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
datapipelines==1.0.7
Django==5.1.1
django-cassiopeia==2.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
flake8==7.1.1
gunicorn==23.0.0
idna==3.10
mccabe==0.7.0
merakicommons==1.0.10
mypy-extensions==1.0.0
networkx==3.3
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
pycodestyle==2.12.1
pyflakes==3.2.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tomli==2.0.2
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
urllib3==2.2.3
